@@ -1,26 +1,41 @@
 {
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
-    "start": "expo start",
+    "dev": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
+    "web": "expo start --web",
     "eject": "expo eject"
   },
   "dependencies": {
-    "expo": "^32.0.0",
-    "firebase": "^6.0.2",
+    "expo": "^35.0.0",
+    "expo-facebook": "^7.0.0",
+    "expo-image-picker": "~7.0.0",
+    "expo-location": "~7.0.0",
+    "expo-permissions": "~7.0.0",
+    "firebase": "^7.2.2",
     "firesql": "^2.0.2",
-    "react": "16.5.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
+    "react": "16.8.3",
+    "react-dom": "16.8.3",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
     "react-native-action-button": "^2.8.5",
+    "react-native-banner-carousel": "^1.0.3",
     "react-native-easy-toast": "^1.2.0",
-    "react-native-elements": "^1.1.0",
-    "react-native-material-bottom-navigation": "^1.0.3",
-    "react-navigation": "^3.9.1",
-    "tcomb-form-native": "^0.6.20"
+    "react-native-elements": "^1.2.7",
+    "react-native-gesture-handler": "1.3.0",
+    "react-native-keyboard-aware-scroll-view": "^0.9.1",
+    "react-native-maps": "~0.25.0",
+    "react-native-open-maps": "^0.3.4",
+    "react-native-reanimated": "1.2.0",
+    "react-native-web": "^0.11.7",
+    "react-navigation": "^4.0.10",
+    "react-navigation-stack": "^1.10.3",
+    "react-navigation-tabs": "^2.5.6",
+    "use-debounce": "^3.1.0",
+    "uuid": "^3.3.3"
   },
   "devDependencies": {
-    "babel-preset-expo": "^5.0.0"
+    "babel-preset-expo": "^7.1.0"
   },
   "private": true
 }
