@@ -1,157 +1,110 @@
-import React, { Component } from "react";
-import { StyleSheet, View, ActivityIndicator } from "react-native";
-import { AirbnbRating, Button, Text, Overlay } from "react-native-elements";
-import Toast, { DURATION } from "react-native-easy-toast";
-
-import t from "tcomb-form-native";
-const Form = t.form.Form;
-import {
-  AddReviewRestaurantStruct,
-  AddReviewRestaurantOptions
-} from "../../forms/AddReviewRestaurant";
+import React, { useState, useRef } from "react";
+import { StyleSheet, View } from "react-native";
+import { AirbnbRating, Button, Input } from "react-native-elements";
+import Toast from "react-native-easy-toast";
+import Loading from "../../components/Loading";
 
 import { firebaseApp } from "../../utils/FireBase";
 import firebase from "firebase/app";
 import "firebase/firestore";
 const db = firebase.firestore(firebaseApp);
 
-export default class AddReviewRestaurant extends Component {
-  constructor(props) {
-    super(props);
+export default function AddReviewRestaurant(props) {
+  const { navigation } = props;
+  const { idRestaurant, setReviewsReload } = navigation.state.params;
+  const [rating, setRating] = useState(null);
+  const [title, setTitle] = useState("");
+  const [review, setReview] = useState("");
+  const [isLoading, setIsLoading] = useState(false);
+  const toastRef = useRef();
 
-    this.state = {
-      loading: false
-    };
-  }
-
-  sendReview = () => {
-    const ratingValue = this.refs.rating.state.position;
-
-    const user = firebase.auth().currentUser;
-
-    this.setState({ loading: true });
-
-    if (ratingValue == 0) {
-      this.refs.toast.show("Tienes que puntuar el restaurante", 1500);
-      this.setState({ loading: false });
+  const addReview = () => {
+    if (rating === null) {
+      toastRef.current.show("No has dado ninguna puntuación");
+    } else if (!title) {
+      toastRef.current.show("El titulo es obligatorio");
+    } else if (!review) {
+      toastRef.current.show("El comnatario es obligatorio");
     } else {
-      const validate = this.refs.addReviewRestaurantForm.getValue();
+      setIsLoading(true);
+      const user = firebase.auth().currentUser;
+      const payload = {
+        idUser: user.uid,
+        avatarUser: user.photoURL,
+        idRestaurant: idRestaurant,
+        title: title,
+        review: review,
+        rating: rating,
+        createAt: new Date()
+      };
 
-      if (!validate) {
-        this.refs.toast.show("Completa el formulario", 1500);
-        this.setState({ loading: false });
-      } else {
-        const user = firebase.auth().currentUser;
-
-        const data = {
-          idUser: user.uid,
-          avatarUser: user.photoURL,
-          idRestaurant: this.props.navigation.state.params.id,
-          title: validate.title,
-          review: validate.review,
-          rating: ratingValue,
-          createAt: new Date()
-        };
-
-        db.collection("reviews")
-          .add(data)
-          .then(() => {
-            const restaurantRef = db
-              .collection("restaurants")
-              .doc(this.props.navigation.state.params.id);
-
-            restaurantRef.get().then(response => {
-              const restaurantData = response.data();
-              const ratingTotal = restaurantData.ratingTotal + ratingValue;
-              const quantityVoting = restaurantData.quantityVoting + 1;
-              const rating = ratingTotal / quantityVoting;
-
-              restaurantRef
-                .update({ rating, ratingTotal, quantityVoting })
-                .then(() => {
-                  this.setState({ loading: false });
-                  this.refs.toast.show(
-                    "Review enviada correctamente",
-                    50,
-                    () => {
-                      this.props.navigation.state.params.loadReviews();
-                      this.props.navigation.goBack();
-                    }
-                  );
-                });
-            });
-          })
-          .catch(() => {
-            this.refs.toast.show(
-              "Error al enviar la review, intentelo mas tarde",
-              1500
-            );
-          });
-      }
+      db.collection("reviews")
+        .add(payload)
+        .then(() => {
+          updateRestaurant();
+        })
+        .catch(() => {
+          toastRef.current.show(
+            "Error al enviar la review, intentelo más tarde"
+          );
+          setIsLoading(false);
+        });
     }
   };
 
-  render() {
-    const { loading } = this.state;
+  const updateRestaurant = () => {
+    const restaurantRef = db.collection("restaurants").doc(idRestaurant);
 
-    return (
-      <View style={styles.viewBody}>
-        <View style={styles.viewRating}>
-          <AirbnbRating
-            ref="rating"
-            count={5}
-            reviews={[
-              "Pésimo",
-              "Deficiente",
-              "Normal",
-              "Muy Bueno",
-              "Excelente"
-            ]}
-            defaultRating={0}
-            size={35}
-          />
-        </View>
+    restaurantRef.get().then(response => {
+      const restaurantData = response.data();
+      const ratingTotal = restaurantData.ratingTotal + rating;
+      const quantityVoting = restaurantData.quantityVoting + 1;
+      const ratingResult = ratingTotal / quantityVoting;
 
-        <View style={styles.formReview}>
-          <Form
-            ref="addReviewRestaurantForm"
-            type={AddReviewRestaurantStruct}
-            options={AddReviewRestaurantOptions}
-          />
-        </View>
+      restaurantRef
+        .update({ rating: ratingResult, ratingTotal, quantityVoting })
+        .then(() => {
+          setIsLoading(false);
+          setReviewsReload(true);
+          navigation.goBack();
+        });
+    });
+  };
 
-        <View style={styles.viewSendReview}>
-          <Button
-            title="Enviar"
-            onPress={() => this.sendReview()}
-            buttonStyle={styles.sendBtnReview}
-          />
-        </View>
-
-        <Overlay
-          overlayStyle={styles.overlayLoading}
-          isVisible={loading}
-          width="auto"
-          height="auto"
-        >
-          <View>
-            <Text style={styles.overlayLoadingText}>Enviando Review</Text>
-            <ActivityIndicator size="large" color="#00a680" />
-          </View>
-        </Overlay>
-
-        <Toast
-          ref="toast"
-          position="bottom"
-          positionValue={320}
-          fadeInDuration={1000}
-          fadeOutDuration={1000}
-          opacity={0.8}
-          textStyle={{ color: "#fff" }}
+  return (
+    <View style={styles.viewBody}>
+      <View style={styles.viewRating}>
+        <AirbnbRating
+          count={5}
+          reviews={["Pésimo", "Deficiente", "Normal", "Muy Bueno", "Excelente"]}
+          defaultRating={0}
+          size={35}
+          onFinishRating={value => setRating(value)}
         />
       </View>
-    );
-  }
+      <View style={styles.formReview}>
+        <Input
+          placeholder="Titulo"
+          containerStyle={styles.input}
+          onChange={e => setTitle(e.nativeEvent.text)}
+        />
+        <Input
+          placeholder="Comentario..."
+          multiline={true}
+          inputContainerStyle={styles.textArea}
+          onChange={e => setReview(e.nativeEvent.text)}
+        />
+        <Button
+          title="Enviar Comentario"
+          onPress={addReview}
+          containerStyle={styles.btnContainer}
+          buttonStyle={styles.btn}
+        />
+      </View>
+      <Toast ref={toastRef} position="center" opacity={0.5} />
+      <Loading isVisible={isLoading} text="Enviando comentario" />
+    </View>
+  );
 }
 
 const styles = StyleSheet.create({
@@ -163,25 +116,28 @@ const styles = StyleSheet.create({
     backgroundColor: "#f2f2f2"
   },
   formReview: {
+    flex: 1,
+    alignItems: "center",
     margin: 10,
     marginTop: 40
   },
-  viewSendReview: {
+  input: {
+    marginBottom: 10
+  },
+  textArea: {
+    height: 150,
+    width: "100%",
+    padding: 0,
+    margin: 0
+  },
+  btnContainer: {
     flex: 1,
     justifyContent: "flex-end",
-    marginBottom: 30,
-    marginLeft: 20,
-    marginRight: 20
+    marginTop: 20,
+    marginBottom: 10,
+    width: "95%"
   },
-  sendBtnReview: {
+  btn: {
     backgroundColor: "#00a680"
-  },
-  overlayLoading: {
-    padding: 20
-  },
-  overlayLoadingText: {
-    color: "#00a680",
-    marginBottom: 20,
-    fontSize: 20
   }
 });
