@@ -1,21 +1,7 @@
 import React from "react";
-import { StyleSheet, Text, View } from "react-native";
-import UserNavigation from "./app/navigations/User";
-
+import Navigation from "./app/navigations/Navigation";
 import { firebaseApp } from "./app/utils/FireBase";
 
-export default class App extends React.Component {
-  render() {
-    return (
-      <View style={styles.container}>
-        <UserNavigation />
-      </View>
-    );
-  }
+export default function App() {
+  return <Navigation />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1
-  }
-});
