@@ -1,4 +1,4 @@
 export const FacebookApi = {
-  application_id: "848749228792334",
+  application_id: "628035101063211",
   permissions: ["public_profile"]
 };
