@@ -1,10 +1,4 @@
-import t from "tcomb-form-native";
-
-export default (formValidation = {
-  email: t.refinement(t.String, value => {
-    return /@/.test(value);
-  }),
-  password: t.refinement(t.String, value => {
-    return value.length >= 6;
-  })
-});
+export function validateEmail(email) {
+  const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+  return re.test(String(email).toLocaleLowerCase());
+}
